feat: Show the performance, vision frame details and give the opportunity to disable vision or mlkit
The FPS, CPU and Memory Usage is now available to compare Vision and MLKit.

Show the current detected word frame counts for Vision and MLKit and give the opportunity to disable the tracking and drawing with a Switch

Add a button to show a random picked detected frame from the vision framework on a detailview. It's a preparation for further features to compare the results, if we use the vision framework as pipeline for the mlkit.