fix: Check the status of running detection tasks and disallow new ones until they are completed
To run a new detection task during the old one is not completed can lead to a crash.
This commit fixed this issues, but leads also to lower detection tasks per second, especially for Vision, if ML Kit is enabled, because both have to complete there tasks until a next pixelbuffer will be inspected.