feat: Add code to collect vision results and present them in a tableView together with a MLKit detection text result
Add a VisionResultListViewController togehter with the TableviewCell, Model and ViewModels.

If the user enable the collect switch, the detected frames from Vision will be collected.
When the user press on the button, the scrapped images will be detected from MLKit and the text results will be presented together with the image.